moved decompression table back into decompression folder
